# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VendorHub - Multivendor Social Media, Sales & Inventory Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/transactions:
    get:
      operationId: listInventoryTransactions
      tags: [inventory]
      summary: List inventory transactions
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: vendorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryTransaction"
    post:
      operationId: createInventoryTransaction
      tags: [inventory]
      summary: Record inventory transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryTransaction"
  /inventory/summary:
    get:
      operationId: getInventorySummary
      tags: [inventory]
      summary: Get inventory summary stats
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventorySummary"

  # ─── ORDERS ───────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List sales orders
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a sales order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/summary:
    get:
      operationId: getOrdersSummary
      tags: [orders]
      summary: Sales summary & revenue stats
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersSummary"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── LEADS ────────────────────────────────────────────────────────────────
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Manually create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/scrape:
    post:
      operationId: scrapeLeads
      tags: [leads]
      summary: Trigger AI lead scraping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadScrapeRequest"
      responses:
        "200":
          description: Scraped leads result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadScrapeResult"
  /leads/stats:
    get:
      operationId: getLeadsStats
      tags: [leads]
      summary: Get lead funnel stats
      parameters:
        - name: vendorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lead stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsStats"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── EMAIL CAMPAIGNS ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    InventoryTransaction:
      type: object
      required: [id, productId, vendorId, type, quantity, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        vendorId:
          type: integer
        type:
          type: string
        quantity:
          type: integer
        notes:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        createdAt:
          type: string
    InventoryTransactionInput:
      type: object
      required: [productId, vendorId, type, quantity]
      properties:
        productId:
          type: integer
        vendorId:
          type: integer
        type:
          type: string
        quantity:
          type: integer
        notes:
          type: string
        reference:
          type: string
    InventorySummary:
      type: object
      required: [totalProducts, totalValue, lowStockCount, outOfStockCount, categories]
      properties:
        totalProducts:
          type: integer
        totalValue:
          type: number
        lowStockCount:
          type: integer
        outOfStockCount:
          type: integer
        categories:
          type: array
          items:
            type: object
            required: [category, count, value]
            properties:
              category:
                type: string
              count:
                type: integer
              value:
                type: number

    # ─── ORDER ──────────────────────────────────────────────────────────────
    Order:
      type: object
      required: [id, vendorId, customerName, customerEmail, status, totalAmount, items, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: ["string", "null"]
        status:
          type: string
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, totalPrice]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
    OrderInput:
      type: object
      required: [vendorId, customerName, customerEmail, items]
      properties:
        vendorId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        notes:
          type: string
        shippingAddress:
          type: string
        items:
          type: array
          items:
            type: object
            required: [productId, quantity, unitPrice]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              unitPrice:
                type: number
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
        shippingAddress:
          type: string
    OrdersSummary:
      type: object
      required: [totalOrders, totalRevenue, pendingOrders, completedOrders, revenueByPeriod]
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: number
        pendingOrders:
          type: integer
        completedOrders:
          type: integer
        revenueByPeriod:
          type: array
          items:
            type: object
            required: [period, revenue, orders]
            properties:
              period:
                type: string
              revenue:
                type: number
              orders:
                type: integer

    # ─── LEAD ───────────────────────────────────────────────────────────────
    Lead:
      type: object
      required: [id, vendorId, name, status, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        score:
          type: ["integer", "null"]
        linkedinUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    LeadInput:
      type: object
      required: [vendorId, name]
      properties:
        vendorId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        industry:
          type: string
        location:
          type: string
        source:
          type: string
        notes:
          type: string
        linkedinUrl:
          type: string
        websiteUrl:
          type: string
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        industry:
          type: string
        location:
          type: string
        status:
          type: string
        notes:
          type: string
        score:
          type: integer
    LeadScrapeRequest:
      type: object
      required: [vendorId, targetIndustry, targetLocation]
      properties:
        vendorId:
          type: integer
        targetIndustry:
          type: string
        targetLocation:
          type: string
        businessType:
          type: string
        keywords:
          type: string
        maxResults:
          type: integer
    LeadScrapeResult:
      type: object
      required: [found, imported, leads]
      properties:
        found:
          type: integer
        imported:
          type: integer
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
    LeadsStats:
      type: object
      required: [totalLeads, newLeads, qualifiedLeads, convertedLeads, byStatus, bySource]
      properties:
        totalLeads:
          type: integer
        newLeads:
          type: integer
        qualifiedLeads:
          type: integer
        convertedLeads:
          type: integer
        byStatus:
          type: array
          items:
            type: object
            required: [status, count]
            properties:
              status:
                type: string
              count:
                type: integer
        bySource:
          type: array
          items:
            type: object
            required: [source, count]
            properties:
              source:
                type: string
              count:
                type: integer

    # ─── EMAIL CAMPAIGN ─────────────────────────────────────────────────────
...